+++ ui/mods/donation_panel/live_game.js
@@ -4,8 +4,14 @@
   paths.sandbox_unit_menu = 'coui://ui/mods/sandbox_unit_menu'
 
   handlers.request_player_names = function(payload) {
-    api.Panel.message(payload[0], payload[1], model.playerData().names)
-    api.panels.options_bar && api.panels.options_bar.message('donation_panel_loading', false)
+    api.Panel.message(payload[0], payload[1],
+      model.playerData().names)
+    api.panels.options_bar && 
+      api.panels.options_bar.message('donation_panel_loading', false)
+  }
+  handlers.request_planet_names = function(payload) {
+    api.Panel.message(payload[0], payload[1],
+      model.celestialViewModels().map(function(planet) {return planet.name()}))
   }
 })()
 
@@ -19,5 +25,9 @@ require(['donation_panel/panel'], function(panel) {
 
   model.playerData.subscribe(function(value) {
     api.Panel.message('donation_panel', 'player_names', value.names)
+  })
+  model.celestialViewModels.subscribe(function(value) {
+    api.Panel.message('donation_panel', 'planet_names',
+      value.map(function(planet) {return planet.name()}))
   })
 })