+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Move fixed elements to bottom
 - Cancel button on each item
 - If a unit already comes in batches, multiply it by increasig the number rather than the instances
+- Attempt to match planet names
 
 ## 1.0.0